New namespace for MsSql and MySql DAO objects.
Cleaned Creation objects by removing validation code and the IValidatable interface. They now are implemented in an independent file.

Adding a NOT NULL property  to a class which already holds records was causing errors in Sql since a value is required. To fix this, a default value is set to the column based in it's default datatype.