Select methods inside Extend class have been deprecated. This is in favor of evolving the query construction based on a Where method. Instead of using Select, switch to Where, which will always return a List of type T. To specify that you only want 1 record, please use QueryOptions.
The Update Stored Procedure has been updated to enable the possibility of updating values to NULL when specified in QueryOptions using the parameter UpdateNulls. This can also be setted when using the Update Extension. If you have been using the library on other projects, please note that for this new feature to work properly, OneData needs to recreate the Update Store Procedure for each of your models, which will happen when you make your next update of each model.